skill_assess/requirements.txt
Django==4.2.9
psycopg2==2.9.9
psycopg2-binary==2.9.9
boto3==1.34.23
django-storages==1.14.2
pdfplumber==0.10.3
google-generativeai==0.3.2
python-dotenv
numpy==1.26.4
keras==3.1.1
opencv-python==4.9.0.80